rename client manager -> customer manager
