+++ src/app/management/management.module.ts
@@ -15,8 +15,8 @@ import { ManagementRoutingGuard } from './management-routing.guard';
 import { ManagementRoutingModule } from './management-routing.module';
 import { ManagementComponent } from './management.component';
 import { ManagementService } from './management.service';
-import { ClientManagerComponent } from './routes/clients/client-manager.component';
-import { ClientManagerService } from './routes/clients/client-manager.service';
+import { CustomerManagerComponent } from './routes/customers/customer-manager.component';
+import { CustomerManagerService } from './routes/customers/customer-manager.service';
 import { ManagementDashboardComponent } from './routes/dashboard/management-dashboard.component';
 import { SellerManagerComponent } from './routes/sellers/seller-manager.component';
 import { SellerManagerService } from './routes/sellers/seller-manager.service';
@@ -45,7 +45,7 @@ const SNACKBAR_DEFAULTS: MatSnackBarConfig = {
     ProductsArrayDialogComponent,
     ManagementDataActionsComponent,
     ManagementDashboardComponent,
-    ClientManagerComponent,
+    CustomerManagerComponent,
     SellerManagerComponent,
     SellerManagerFormDialogComponent,
     ProductManagerComponent,
@@ -63,7 +63,7 @@ const SNACKBAR_DEFAULTS: MatSnackBarConfig = {
     { provide: MAT_SNACK_BAR_DEFAULT_OPTIONS, useValue: SNACKBAR_DEFAULTS },
     ManagementRoutingGuard,
     ManagementService,
-    ClientManagerService,
+    CustomerManagerService,
     SellerManagerService,
     ProductManagerService,
     UserManagerService,
