+++ src/app/management/routes/customers/customer-manager.component.ts
@@ -8,24 +8,24 @@ import { ActivatedRoute } from '@angular/router';
 import { Observable } from 'rxjs';
 import { Customer } from 'src/app/models/entities/Customer';
 import { DataManagerComponentDirective } from '../data-manager.component-directive';
-import { ClientManagerService } from './client-manager.service';
+import { CustomerManagerService } from './customer-manager.service';
 
 @Component({
-  selector: 'app-client-manager',
-  templateUrl: './client-manager.component.html',
+  selector: 'app-customer-manager',
+  templateUrl: './customer-manager.component.html',
   styleUrls: [
     '../data-manager.styles.css',
-    './client-manager.component.css'
+    './customer-manager.component.css'
   ]
 })
-export class ClientManagerComponent
+export class CustomerManagerComponent
   extends DataManagerComponentDirective<Customer>
   implements OnInit {
 
   public tableColumns: string[] = [ 'name', 'idCard' ];
 
   constructor(
-    protected service: ClientManagerService,
+    protected service: CustomerManagerService,
     protected route: ActivatedRoute
   ) {
     super();
